backend/requirements.txt
     Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
google-auth==2.27.0
python-dotenv==1.0.0
bcrypt==4.1.2
SQLAlchemy==2.0.25
Flask-Migrate==3.1.0
python-jose==3.3.0
requests==2.31.0
gunicorn==21.2.0
